Much better

+++ DennisRitchie.c
@@ -1,29 +1,29 @@
 #include <stdio.h>
 
-int add(int num1, int num2);
-int subtract(int num1, int num2);
-int divide(int num1, int num2);
-int multiply(int num1, int num2);
+float add(float num1, float num2);
+float subtract(float num1, float num2);
+float divide(float num1, float num2);
+float multiply(float num1, float num2);
 
 int main(){
-    int a = 5;
-    int b = 5;
-    printf("Sum : %d\n", add(a, b));
-    printf("Difference : %d\n", subtract(a, b));
-    printf("Quotient : %d\n", divide(a, b));
-    printf("Product : %d\n", multiply(a, b));
+    float a = 5;
+    float b = 5;
+    printf("Sum : %.2f\n", add(a, b));
+    printf("Difference : %.2f\n", subtract(a, b));
+    printf("Quotient : %.2f\n", divide(a, b));
+    printf("Product : %.2f\n", multiply(a, b));
     return 0;
 }
 
-int add(int num1, int num2){
+float add(float num1, float num2){
     return num1 + num2;
 }
-int subtract(int num1, int num2){
+float subtract(float num1, float num2){
     return num1 - num2;
 }
-int divide(int num1, int num2){
+float divide(float num1, float num2){
     return num1 / num2;
 }
-int multiply(int num1, int num2){
+float multiply(float num1, float num2){
     return num1 * num2;
 }
